requirements.txt
absl-py~=0.7
numpy~=1.14
six~=1.10
scipy
